gw reporting startdate by trust id
